+++ fabric/src/main/java/io/github/sakurawald/fuji/core/auxiliary/minecraft/RegistryHelper.java
@@ -13,7 +13,6 @@ import net.minecraft.registry.Registries;
 import net.minecraft.registry.Registry;
 import net.minecraft.registry.RegistryEntryLookup;
 import net.minecraft.registry.RegistryKey;
-import net.minecraft.registry.RegistryKeys;
 import net.minecraft.registry.RegistryWrapper;
 import net.minecraft.registry.entry.RegistryEntry;
 import net.minecraft.util.Identifier;
@@ -69,7 +68,7 @@ public class RegistryHelper {
 
         #if MC_VER <= MC_1_20_4
         MessageType messageTypeObj = parameters.type();
-        messageTypeIdString = RegistryHelper.findRegistryKey(RegistryKeys.MESSAGE_TYPE, messageTypeObj)
+        messageTypeIdString = RegistryHelper.findRegistryKey(net.minecraft.registry.RegistryKeys.MESSAGE_TYPE, messageTypeObj)
             .map(RegistryHelper::getIdAsString)
             .orElseThrow(() -> new IllegalStateException("Failed to find the RegistryKey for MessageType %s".formatted(messageTypeObj)));
         #elif MC_VER > MC_1_20_4
@@ -137,6 +136,7 @@ public class RegistryHelper {
         }
     }
 
+    @SuppressWarnings("unused")
     public static <T> Optional<RegistryKey<T>> findRegistryKey(@NotNull RegistryKey<? extends Registry<? extends T>> registrySpecifier, @NotNull T registryValue) {
         return RegistryHelper
             .getRegistry(registrySpecifier)